refactor : move button to component folder
